+++ Libplanet.Tests/Blockchain/NullPolicy.cs
@@ -26,6 +26,8 @@ namespace Libplanet.Tests.Blockchain
 
         public IAction BlockAction => null;
 
+        public int MinTransactionsPerBlock => 0;
+
         public int GetMaxTransactionsPerBlock(long index) => int.MaxValue;
 
         public bool DoesTransactionFollowsPolicy(